ui_inference_container/pyproject.toml
[project]
name = "ui-inference-container"
version = "0.1.0"
description = ""
authors = [
    {name = "Marcel Knauf",email = "marcel.kn1994@hotmail.de"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy (>=2.2.6,<3.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "mlflow (>=2.22.0,<3.0.0)",
    "python-dotenv (>=0.21.0,<1.0.0)",
    "dash (>=3.0.4,<4.0.0)",
    "dash-bootstrap-components (>=2.0.3,<3.0.0)",
    "palmerpenguins (>=0.1.4,<0.2.0)",
    "setuptools (>=80.8.0,<81.0.0)",
    "boto3 (>=1.28.0,<2.0.0)", # Added boto3
    "pytest (>=8.3.5,<9.0.0)",
    "httpx (>=0.28.1,<0.29.0)"
]

[tool.poetry]
name = "pinguin-ui-api"
version = "0.1.0"
description = "UI und Inferenz für Pinguin-Klassifikator"
authors = ["Marcel Knauf <marcel.kn1994@hotmail.de>"]
packages = [{include = "ui_app"}]

[tool.poetry.dependencies]
python = "^3.10"
dash = ">=3.0.4,<4.0.0"  # Aligned with [project]
pandas = ">=2.2.3,<3.0.0" # Aligned with [project]
mlflow = ">=2.22.0,<3.0.0" # Aligned with [project]
python-dotenv = ">=0.21.0,<1.0.0" # Aligned with [project]
scikit-learn = ">=1.6.1,<2.0.0" # Aligned with [project]
requests = "^2.31.0"
sqlalchemy = "^2.0.29"
psycopg2-binary = "^2.9.9"
boto3 = ">=1.28.0,<2.0.0" # Assuming similar boto3 requirement as training, adjust if not
numpy = ">=2.2.6,<3.0.0" # Added from [project]
dash-bootstrap-components = ">=2.0.3,<3.0.0" # Added from [project]
palmerpenguins = ">=0.1.4,<0.2.0" # Added from [project]
setuptools = ">=80.8.0,<81.0.0" # Added from [project]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
